Skip tests incompatible with 2.7 and earlier
Not a lot of energy around making these 2.7 compat as is an older
ansible, not going to block this work any further waiting to see what we
might want to do around these.

If it becomes a priority to add more coverage for 2.7 in future we can
start by looking at these tests.
